% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_training_test.R
\name{combine_training_test}
\alias{combine_training_test}
\title{Combining training and test data}
\usage{
combine_training_test(object)
}
\arguments{
\item{object}{An object of class \code{Mcomp} from either the Mcomp or Tcomp
packages.}
}
\value{
A list of time series
}
\description{
For each time series, the training data and test data are combined into
a single \code{ts} object.
}
\examples{
m3 <- combine_training_test(M3)
}
