% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gefcom2012_temp}
\alias{gefcom2012_temp}
\title{GEFCOM2012 temperature data}
\format{A list of time series data, each of class \code{msts}.}
\source{
\url{http://www.drhongtao.com/gefcom/2012}.
}
\usage{
gefcom2012_temp
}
\description{
Data from the GEFCOM2012 forecasting competition, comprising 11 time series containing
hourly temperature data from 11 weather stations in the United States.
}
\examples{
plot(gefcom2012_temp[[1]])
}
\keyword{datasets}
