% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nn3}
\alias{nn3}
\title{NN3 data}
\format{A list of time series data}
\source{
\url{http://www.neural-forecasting-competition.com/NN3}.
}
\usage{
nn3
}
\description{
Data from the NN3 forecasting competition, comprising 111 monthly time series.
Training and test data are combined. In the competition, the last 18 months 
were used as test data. Time series NN3-101 to NN3-111 made up the reduced data
set from the competition.
}
\examples{
plot(nn3[[1]])

}
\keyword{datasets}
