% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nn5}
\alias{nn5}
\title{NN5 data}
\format{A list of time series data, each of class \code{msts}.}
\source{
\url{http://www.neural-forecasting-competition.com/NN5}.
}
\usage{
nn5
}
\description{
Data from the NN5 forecasting competition, comprising 111 daily time series.
Training and test data are combined. In the competition, the last 56 days 
were used as test data. Time series NN5-101 to NN5-111 made up the reduced data
set from the competition.
}
\examples{
plot(nn5[[1]])

}
\keyword{datasets}
