% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nngc1}
\alias{nngc1}
\title{NNGC1 data}
\format{A list of time series data, each of class \code{ts} or \code{msts}.}
\source{
\url{http://www.neural-forecasting-competition.com/}.
}
\usage{
nngc1
}
\description{
Data from the NNGC1 forecasting competition, comprising 11 annual time series, 11
quarterly time series, 11 monthly time series, 11 weekly time series, 11 daily
time series and 11 hourly time series.  Only training data are provided.
}
\examples{
plot(nngc1[[1]])

}
\keyword{datasets}
