% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tscompdata.R
\docType{package}
\name{tscompdata}
\alias{tscompdata}
\alias{tscompdata-package}
\title{Time Series Competition Data
 
The tscompdata package provides time series data from several forecasting competitions.}
\description{
Time Series Competition Data
 
The tscompdata package provides time series data from several forecasting competitions.
}
