% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gefcom2012_load}
\alias{gefcom2012_load}
\title{GEFCOM2012 load data}
\format{A list of time series data, each of class \code{msts}.}
\source{
\url{http://www.drhongtao.com/gefcom/2012}.
}
\usage{
gefcom2012_load
}
\description{
Data from the GEFCOM2012 forecasting competition, comprising 20 time series containing
hourly load data from 20 zones in the United States. Only training data are provided.
The missing data in each series formed the test sets.
}
\examples{
plot(gefcom2012_load[[1]])
}
\keyword{datasets}
