% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gefcom2012_wp}
\alias{gefcom2012_wp}
\title{GEFCOM2012 wind power data}
\format{A list of time series data, each of class \code{msts}.}
\source{
\url{https://www.kaggle.com/c/GEF2012-wind-forecasting/data},
\url{http://www.drhongtao.com/gefcom/2012}.
}
\usage{
gefcom2012_wp
}
\description{
Data from the GEFCOM2012 forecasting competition, comprising 7 hourly time series containing
wind power data from 7 wind farms. Only training data are provided.
The missing data in each series formed the test sets.
}
\examples{
plot(gefcom2012_wp[[1]])
}
\keyword{datasets}
